requirements.txt
biopython==1.86
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dash==3.3.0
dash-bootstrap-components==2.0.4
dash-bootstrap-templates==2.1.0
dash-iconify==0.1.2
dash_mantine_components==2.4.1
Flask==3.1.2
fonttools==4.61.1
greenlet==3.3.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mysql-connector-python==9.5.0
narwhals==2.14.0
nest-asyncio==1.6.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
requests==2.32.5
retrying==1.4.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.45
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
Werkzeug==3.1.4
zipp==3.23.0
